tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/csskeyframesrule_findrule_no_match.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - findRule on CSSKeyframesRule, no match</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#dom-interfaces-">
   <meta name="flags" content="dom">
   <meta name="assert" content="If no rule with matching key exists, findRule
   returns a null value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes test {
         from { width: 100px; }
         to { width: 200px; }
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = false;
         try {
            var testKeyframes = document.styleSheets[0].cssRules[0];
            passed = (testKeyframes.findRule('50%') === null);
         } catch(e) {
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
</body>
</html>
